RowRunner를 이용해 Row 리팩토링 Row에 runWithPrint 메서드 추가
