+++ 이현희/untitled/src/test/java/RowRunnerTest.java
@@ -7,24 +7,13 @@ public class RowRunnerTest {
     @Test
     public void 실행하기() {
         //given
-        RowRunner rowRunner = new RowRunner();
+        int size = 5;
+        int[] board = new int[size];
+        RowRunner rowRunner = new RowRunner(board, size);
 
         //when
         int startPos = 3;
         int ret = rowRunner.run(3);
-
-        //then
-        assertEquals(ret, 3);
-    }
-
-    @Test
-    public void 출력하며실행하기() {
-        //given
-        RowRunner rowRunner = new RowRunner();
-
-        //when
-        int startPos = 3;
-        int ret = rowRunner.runWithPrint(3);
 
         //then
         assertEquals(ret, 3);
